Adjusted to noteful-server
I changed the prop-types of folderId, and noteId to validate for number rather than string since I updated that id values.

The endpoints are working as expected after creating the CRUD service objects, and adding the routers which are then added to app.js

I need to add testing for both the folder, and notes endpoints. Finished up to the point of folders endpoints for GET /api/folders endpoint.

Action needed: The react client, and server are still not in sync. The GET requests work perfectly fine. I receive an error, however, when trying to submit a POST, DELETE, or PATCH request.